deleted nav-bar.component.js, replaced it with NAV_BAR_DATA.js, restructured home-page to render nav-bar by mapping NAV_BAR_DATA and placing nested routed in the Col. to form layout
