+++ client/src/components/home-page/home-page.component.js
@@ -1,12 +1,20 @@
 import TopSection from "../top-section/top-section.component";
-import NavBar from "../nav-bar/nav-bar.component";
-import RecipeCard from "../recipe-card/recipe-card.component";
-import ViewRecipe from "../view-recipe/view-recipe.component";
-import NavBarPages from "../nav-bar-pages/nav-bar-pages";
-import SearchBar from "../search-bar/search-bar.component";
 import { Container, Row, Col } from "react-bootstrap";
+import { Outlet } from "react-router-dom";
+import { NAV_BAR_DATA } from "../nav-bar/nav-bar-data.js";
+import { NavLink, useNavigate } from "react-router-dom";
+import { useContext } from "react";
+import { UserContext } from "../../contexts/user.context";
+import { ReactComponent as ProfileLogo } from "../../assets/profile.svg";
 
 const HomePage = () => {
+  const { currentUser, setCurrentUser } = useContext(UserContext);
+  const navigate = useNavigate();
+
+  const handleLogout = () => {
+    setCurrentUser({});
+    navigate("/");
+  };
   return (
     <Container fluid style={{ height: "100%", width: "100vw" }}>
       <Row
@@ -21,26 +29,32 @@ const HomePage = () => {
           <TopSection />
         </Col>
       </Row>
-      <Row
-        style={{
-          height: "5%",
-          width: "100vw",
-        }}
-      >
-        <Col className="d-flex justify-content-center">
-          <SearchBar />
-        </Col>
-      </Row>
       <Row className="h-100">
-        <Col className="col-2 text-center">
-          <NavBar />
+        <Col className="col-2 d-flex flex-column h3 text-center mt-5">
+          <ProfileLogo className="mx-auto" />
+          {NAV_BAR_DATA.map((item, index) => {
+            return (
+              <div key={index} className="py-4 my-3 bg-secondary">
+                <NavLink
+                  to={item.path}
+                  className="text-white text-decoration-none"
+                >
+                  {item.title}
+                </NavLink>
+              </div>
+            );
+          })}
+          <button
+            type="button"
+            className="py-4 my-3 bg-secondary h3 text-white"
+            onClick={handleLogout}
+          >
+            Log Out
+          </button>
         </Col>
         <Col className="col">
-          <NavBarPages />
+          <Outlet />
         </Col>
-        {/* {<RecipeCard />} */}
-        {/* && <ViewRecipe /> */}
-        {/* && <CreateRecipe /> */}
       </Row>
     </Container>
   );
