--- client/src/components/nav-bar-pages/nav-bar-pages.js
@@ -1,9 +1,0 @@
-const NavBarPages = () => {
-  return (
-    <>
-      <div>hi</div>
-    </>
-  );
-};
-
-export default NavBarPages;
